+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change all the `_outputs.` to `_outputs__`
 - Disable cdc on output tables.
 - Remove `-` from the uuid of the component.
+- Add _execute_select and filter in the Query class.
+
 
 
 #### New Features & Functionality